+++ src/proto_helpers.rs
@@ -54,10 +54,10 @@ pub fn create_stream_header_field(key: String, value: Vec<u8>) -> ChunkBidi {
     }
 }
 
-pub fn create_data_chunk(content_length: Option<i64>, chunk: Vec<u8>) -> ChunkBidi {
+pub fn create_data_chunk(content_length: Option<usize>, chunk: Vec<u8>) -> ChunkBidi {
     let header = content_length.map(|len| StreamHeader {
         name: consts::DIME_FIELD_NAME.to_owned(),
-        content_length: len,
+        content_length: len as i64,
     });
 
     let data_chunk = Chunk {
